title-html: Allow titles to contain HTML content
Since the input format must be XML-compliant, the characters are escaped
as entities in the title, and unescaped upon output. Ideally this would
be handled by creating a dedicated `<title>` node, but givent that the
course has already started, I prefered keeping the changes minimal, to
avoid having to update the XML content for the whole course.